feral dps WIP Removed misc SoD stuff Added PVP and PVE set bonuses Added P2 gear sets Added P2 APL rotation
