+++ sim/druid/item_sets_pve.go
@@ -1,0 +1,135 @@
+package druid
+
+import (
+	"github.com/wowsims/classic/sim/core"
+	"github.com/wowsims/classic/sim/core/stats"
+)
+
+var ItemSetFeralheartRaiment = core.NewItemSet(core.ItemSet{
+	Name: "Feralheart Raiment",
+	Bonuses: map[int32]core.ApplyEffect{
+		// (2) Set : +8 All Resistances.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddResistances(8)
+		},
+		// (4) Set : When struck in combat has a chance of returning 300 mana, 10 rage, or 40 energy to the wearer. (Proc chance: 2%)
+		4: func(agent core.Agent) {
+			/* TODO: check this sod stuff against vanilla
+			c := agent.GetCharacter()
+			actionID := core.ActionID{SpellID: 450608}
+			manaMetrics := c.NewManaMetrics(actionID)
+			energyMetrics := c.NewEnergyMetrics(actionID)
+			rageMetrics := c.NewRageMetrics(actionID)
+
+			core.MakeProcTriggerAura(&c.Unit, core.ProcTrigger{
+				Name:       "S03 - Druid Energize Trigger - Wildheart Raiment (Mana)",
+				Callback:   core.CallbackOnCastComplete,
+				ProcMask:   core.ProcMaskSpellDamage | core.ProcMaskSpellHealing,
+				ProcChance: 0.02,
+				Handler: func(sim *core.Simulation, spell *core.Spell, _ *core.SpellResult) {
+					c.AddMana(sim, 300, manaMetrics)
+				},
+			})
+			core.MakeProcTriggerAura(&c.Unit, core.ProcTrigger{
+				Name:       "S03 - Druid Energize Trigger - Wildheart Raiment (Energy)",
+				Callback:   core.CallbackOnSpellHitDealt,
+				Outcome:    core.OutcomeLanded,
+				ProcMask:   core.ProcMaskMeleeWhiteHit,
+				ProcChance: 0.06,
+				Handler: func(sim *core.Simulation, spell *core.Spell, _ *core.SpellResult) {
+					if c.HasEnergyBar() {
+						c.AddEnergy(sim, 40, energyMetrics)
+					}
+				},
+			})
+			core.MakeProcTriggerAura(&c.Unit, core.ProcTrigger{
+				Name:       "S03 - Druid Energize Trigger - Wildheart Raiment (Rage)",
+				Callback:   core.CallbackOnSpellHitTaken,
+				ProcMask:   core.ProcMaskMelee,
+				ProcChance: 0.03,
+				Handler: func(sim *core.Simulation, spell *core.Spell, _ *core.SpellResult) {
+					if c.HasRageBar() {
+						c.AddRage(sim, 10, rageMetrics)
+					}
+				},
+			}) */
+		},
+		// (6) Set : Increases damage and healing done by magical spells and effects by up to 15.
+		// (6) Set : +26 Attack Power.
+		6: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellPower, 15)
+			c.AddStat(stats.AttackPower, 26)
+		},
+		// (8) Set : +200 Armor.
+		8: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Armor, 200)
+		},
+	},
+})
+
+var ItemSetCenarionRaiment = core.NewItemSet(core.ItemSet{
+	Name: "Cenarion Raiment",
+	Bonuses: map[int32]core.ApplyEffect{
+		// (3) Set : Damage dealt by Thorns increased by 4 and duration increased by 50%.
+		3: func(agent core.Agent) {
+			// Nothing to do
+		},
+		// (5) Set : Improves your chance to get a critical strike with spells by 2%.
+		5: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.SpellCrit, 2*core.SpellCritRatingPerCritChance)
+		},
+		// (8) Set : Reduces the cooldown of your Tranquility and Hurricane spells by 50%.
+		8: func(agent core.Agent) {
+			// Nothing to do
+		},
+	},
+})
+
+var ItemSetWildheartRaiment = core.NewItemSet(core.ItemSet{
+	Name: "Wildheart Raiment",
+	Bonuses: map[int32]core.ApplyEffect{
+		// (2) Set : +200 Armor.
+		2: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.Armor, 200)
+		},
+		// (4) Set : +26 Attack Power.
+		// (4) Set : Increases damage and healing done by magical spells and effects by up to 15.
+		4: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddStat(stats.AttackPower, 26)
+			c.AddStat(stats.SpellPower, 15)
+		},
+		// (6) Set : When struck in combat has a chance of returning 300 mana, 10 rage, or 40 energy to the wearer. (Proc chance: 2%)
+		6: func(agent core.Agent) {
+			// TODO
+		},
+		// (8) Set : +8 All Resistances.
+		8: func(agent core.Agent) {
+			c := agent.GetCharacter()
+			c.AddResistances(8)
+		},
+	},
+})
+
+var ItemSetStormrageRaiment = core.NewItemSet(core.ItemSet{
+	Name: "Stormrage Raiment",
+	Bonuses: map[int32]core.ApplyEffect{
+		// (3) Set : Allows 15% of your Mana regeneration to continue while casting.
+		3: func(agent core.Agent) {
+			// TODO
+		},
+		// (5) Set : Reduces the casting time of your Regrowth spell by 0.2 sec.
+		5: func(agent core.Agent) {
+			// Nothing to do.
+		},
+		// (8) Set : Increases the duration of your Rejuvenation spell by 3 sec.
+		8: func(agent core.Agent) {
+			// Nothing to do.
+		},
+	},
+})
