+++ ui/feral_druid/apls/p2.apl.json
@@ -1,0 +1,16 @@
+{"type": "TypeAPL",
+  "prepullActions": [
+    {"action":{"castSpell":{"spellId":{"spellId":17392,"rank":4}}},"doAtValue":{"const":{"val":"-1.5"}}}
+  ],
+  "priorityList": [
+    {"action":{"condition":{"and":{"vals":[{"cmp":{"op":"OpLe","lhs":{"auraRemainingTime":{"sourceUnit":{"type":"CurrentTarget"},"auraId":{"spellId":17392,"rank":4}}},"rhs":{"const":{"val":"1"}}}},{"cmp":{"op":"OpLe","lhs":{"auraRemainingTime":{"sourceUnit":{"type":"CurrentTarget"},"auraId":{"spellId":9907,"rank":4}}},"rhs":{"const":{"val":"1"}}}}]}},"castSpell":{"spellId":{"spellId":17392,"rank":4}}}},
+    {"action":{"condition":{"and":{"vals":[{"not":{"val":{"auraIsActive":{"auraId":{"spellId":13494}}}}},{"cmp":{"op":"OpLt","lhs":{"currentTime":{}},"rhs":{"const":{"val":"89"}}}}]}},"castSpell":{"spellId":{"spellId":13494}}}},
+    {"action":{"castSpell":{"spellId":{"spellId":9830,"rank":5}}}},
+    {"action":{"condition":{"and":{"vals":[{"auraIsKnown":{"auraId":{"spellId":17061,"rank":5}}},{"cmp":{"op":"OpGe","lhs":{"currentMana":{}},"rhs":{"spellCurrentCost":{"spellId":{"spellId":768}}}}},{"cmp":{"op":"OpGt","lhs":{"timeToEnergyTick":{}},"rhs":{"const":{"val":"1.02"}}}},{"cmp":{"op":"OpEq","lhs":{"currentComboPoints":{}},"rhs":{"const":{"val":"5"}}}}]}},"castSpell":{"spellId":{"spellId":31018}}}},
+    {"action":{"condition":{"and":{"vals":[{"auraIsKnown":{"auraId":{"spellId":17061,"rank":5}}},{"gcdIsReady":{}},{"cmp":{"op":"OpGe","lhs":{"currentMana":{}},"rhs":{"spellCurrentCost":{"spellId":{"spellId":768}}}}},{"cmp":{"op":"OpLt","lhs":{"currentEnergy":{}},"rhs":{"math":{"op":"OpSub","lhs":{"spellCurrentCost":{"spellId":{"spellId":9830,"rank":5}}},"rhs":{"const":{"val":"20.2"}}}}}}]}},"cancelAura":{"auraId":{"spellId":768}}}},
+    {"action":{"condition":{"and":{"vals":[{"not":{"val":{"auraIsActive":{"auraId":{"spellId":768}}}}},{"cmp":{"op":"OpLe","lhs":{"math":{"op":"OpAdd","lhs":{"currentMana":{}},"rhs":{"const":{"val":"1500"}}}},"rhs":{"math":{"op":"OpDiv","lhs":{"currentMana":{}},"rhs":{"currentManaPercent":{}}}}}}]}},"castSpell":{"spellId":{"itemId":12662}}}},
+    {"action":{"condition":{"and":{"vals":[{"not":{"val":{"auraIsActive":{"auraId":{"spellId":768}}}}},{"cmp":{"op":"OpLe","lhs":{"math":{"op":"OpAdd","lhs":{"currentMana":{}},"rhs":{"const":{"val":"2250"}}}},"rhs":{"math":{"op":"OpDiv","lhs":{"currentMana":{}},"rhs":{"currentManaPercent":{}}}}}}]}},"castSpell":{"spellId":{"otherId":"OtherActionPotion"}}}},
+    {"action":{"condition":{"and":{"vals":[{"not":{"val":{"auraIsActive":{"auraId":{"spellId":768}}}}},{"cmp":{"op":"OpLe","lhs":{"currentManaPercent":{}},"rhs":{"const":{"val":"40%"}}}},{"cmp":{"op":"OpGt","lhs":{"currentMana":{}},"rhs":{"math":{"op":"OpAdd","lhs":{"spellCurrentCost":{"spellId":{"spellId":29166}}},"rhs":{"spellCurrentCost":{"spellId":{"spellId":768}}}}}}},{"not":{"val":{"or":{"vals":[{"and":{"vals":[{"spellIsKnown":{"spellId":{"itemId":12662}}},{"spellIsReady":{"spellId":{"itemId":12662}}}]}},{"and":{"vals":[{"spellIsKnown":{"spellId":{"otherId":"OtherActionPotion"}}},{"spellIsReady":{"spellId":{"otherId":"OtherActionPotion"}}}]}}]}}}}]}},"castSpell":{"spellId":{"spellId":29166}}}},
+    {"action":{"condition":{"not":{"val":{"auraIsActive":{"auraId":{"spellId":768}}}}},"castSpell":{"spellId":{"spellId":768}}}}
+  ] 
+}
