+++ ui/feral_druid/gear_sets/p2.bis.gear.json
@@ -1,0 +1,19 @@
+{"items": [
+    {"id":8345,"enchant":1508},
+    {"id":18404},
+    {"id":12927},
+    {"id":18541,"enchant":849},
+    {"id":22877,"enchant":1891},
+    {"id":19587,"enchant":1885},
+    {"id":15063,"enchant":1887},
+    {"id":13252},
+    {"id":15062,"enchant":1508},
+    {"id":22852,"enchant":1887},
+    {"id":17063},
+    {"id":19325},
+    {"id":13965},
+    {"id":11815},
+    {"id":9449,"enchant":34},
+    {},
+    {}
+]}
