+++ ui/feral_druid/gear_sets/p2.pre-bis.gear.json
@@ -1,0 +1,19 @@
+{"items": [
+    {"id":8345,"enchant":1508},
+    {"id":15411},
+    {"id":12927},
+    {"id":13340,"enchant":849},
+    {"id":14637,"enchant":1891},
+    {"id":13120,"enchant":1885},
+    {"id":15063,"enchant":1887},
+    {"id":13252},
+    {"id":15062,"enchant":1508},
+    {"id":12553,"enchant":1887},
+    {"id":18500},
+    {"id":18500},
+    {"id":13965},
+    {"id":19120},
+    {"id":9449,"enchant":34},
+    {},
+    {}
+]}
