+++ ui/index.html
@@ -122,13 +122,13 @@
 										</a>
 									</li>
 									<li>
-										<a href="#" class="sim-link text-druid">
+										<a href="/classic/feral_druid/" class="sim-link text-druid">
 											<div class="sim-link-content">
 												<img src="https://wow.zamimg.com/images/wow/icons/large/ability_druid_catform.jpg" class="sim-link-icon" />
 												<div class="d-flex flex-column">
 													<span class="sim-link-label">Druid</span>
 													<span class="sim-link-title">Feral DPS</span>
-													<span class="launch-status-label text-brand">Phase 1 - Unreleased</span>
+													<span class="launch-status-label text-brand">Phase 1 - Alpha</span>
 												</div>
 											</div>
 										</a>
